Move all files into thanos package
