+++ src/main/java/thanos/commands/ByeCommand.java
@@ -1,8 +1,8 @@
-package commands;
+package thanos.commands;
 
-import exceptions.InvalidCommandException;
-import tasks.TaskList;
-import ui.Ui;
+import thanos.exceptions.InvalidCommandException;
+import thanos.tasks.TaskList;
+import thanos.ui.Ui;
 
 public class ByeCommand extends Command {
     @Override
