+++ src/main/java/thanos/commands/Command.java
@@ -1,8 +1,8 @@
-package commands;
+package thanos.commands;
 
-import exceptions.InvalidCommandException;
-import tasks.TaskList;
-import ui.Ui;
+import thanos.exceptions.InvalidCommandException;
+import thanos.tasks.TaskList;
+import thanos.ui.Ui;
 
 public abstract class Command {
     public abstract void execute(TaskList taskList, Ui ui) throws InvalidCommandException;
