+++ src/main/java/thanos/tasks/Todo.java
@@ -1,4 +1,4 @@
-package tasks;
+package thanos.tasks;
 
 public class Todo extends Task {
     public Todo(String description) {
